requirements.txt
requests
beautifulsoup4
lxml
flask
flask-cors
langchain
langchain-community
langchain-groq
python-dotenv
fastapi
uvicorn
jinja2
playwright
python-multipart
aiofiles
sqlalchemy
sqlite3
nltk